refactor(mvc): moved all services in `/app/services`
