+++ app/interfaces/services/service.py
@@ -1,0 +1,5 @@
+import typing as t
+
+
+class IService(t.Protocol):
+    ...
